Explain rollback philosophy (#5)
We've had a few issues (#3 and #4) opened related to supporting
ActiveRecord's automatic migration rollback capability. That made me
realize we hadn't properly documented our approach to this concern.

Additionally I've added proper links to our extended blog post on
migrations without downtime to provide the in-depth reasoning behind
many of the choices we've made in the gem.