Fix ActiveRecord migration compatibility inheritance, part 2
ActiveRecord implements version backwards compatibility for migrations
via per-migration version subclasses. This is rather at odds with our
strategy of prepending overrides, since it's not obvious where to put
our modules in the ancestor chain. For example, if we put it at the
bottom, only migrations of the oldest compatibility version will
benefit, but if we prepend on every version subclass the modules will be
in ancestors multiple times for the lower version subclasses.

Since most of migration's functionality is implemented as instance
methods we can't rely on delegating to the superclass (though
confusingly the migration class makes this appear to work), and we have
to maintain per-version implementation details anyway (so calling one
class up with the overrides prepended on every version can end up
unintentionally triggering our `raise` guard methods).

To be able to maintain our goals of:
1. Preventing non-explicit uses of original/unsafe implementations, and
2. Wrapping (instead of re-implementing) the unsafe implementations
we need to be able to access the original implementation while still not
having the original method directly available.

Further complicating things much of `ActiveRecord::Migration`'s
functionality isn't actually implemented on `Migration` directly;
instead `ActiveRecord::Migration` proxies it to the connection via
`method_missing`.

One possible solution would be to implement our own versioned subclasses
and require migrations to inherit from those classes. But that would
break our design goal of making it difficult to use the original unsafe
implementations without communicating explicit intent to do so.

The only implementation that seems to satisfy all of the goals above is
to apply the overrides on a per-application-migration-subclass level and
implement some reasonable complicated method dispatch logic for finding
the original implementation we want to wrap with our safety guards.

Co-authored-by: Celeen <celeen@users.noreply.github.com>