Check for dependent indexes before dropping columns (#18)
We generally have two major types of concerns we care about for the
safe/unsafe distinction:
- Database safety (e.g., long-held locks)
- Application safety (e.g., dropping columns the app uses)

Currently we check for both of those concerns in the `safe_*` method
variants, and the `unsafe_*` variants are always a direct dispatch to
the native ActiveRecord migration method.

But it seems there's a class of problems that we can't handle easily
without breaking that design rule a bit. For example, dropping a column
is unsafe from an application perspective, so we want to call it out
explicitly with an `unsafe_` prefix to give it extra visibility (someone
needs to go confirm the application is rollback safe without this
column present) -- that means there is no `safe_remove_column`. However
there are other conditions we'd like to assert before dropping a column:
for example, dropping an unused column that's contained in one or more
indexes may be safe from an application perspective, but the cascading
drop of the index won't use the `CONCURRENT` operation form and is
therefore unsafe from a database perspective.

In this changeset I've relaxed the "`unsafe_*` variants always directly
dispatch to the native AR method" rule a bit and introduced the concept
of "dependent object checks". For this first iteration I've decided to
only handle the `remove_column` case where at least one index depends on
the column being removed, but I've tried to ensure it's reasonably
extensible.