Disallow `force: true` when creating table (#28)
The `force: true` option to `ActiveRecord::Migration#create_table`
causes the table to be dropped if it already exists. This is unsafe by
definition, and should instead require an explicit `drop_table` instead
-- even when calling the `unsafe_create_table` variant.

We always disallow the option to `safe_create_table`, but to ensure we
don't break backwards compatibility for `unsafe_create_table`, I've
configured the default for this new feature to continue to allow the
old behavior -- making it opt-in for now, but with the expectation that
we'll flip the default in the next major version.