Fix safe_change_column_default to support DROP DEFAULT case
The original implementation was re-rolled as opposed to just wrapping
the `change_column_default` implementation from ActiveRecord for
somewhat unknown reasons. Then on top of that we'd added support for
procs (ActiveRecord's workaround for setting default expressions rather
than hard-coded values), but the implementation of that was also
(we think incorrectly) under the assumption that up-front evaluation of
non-proc'd expressions needed to be supported. This wouldn't have worked
with all datatypes anyway, and there's a more explicit way to do it
anyway (pass a _quoted_ expression in a proc), so we've removed that
support. That also fixed the inability to drop the default by setting it
to NULL, since that "feature" had been incorrectly coercing NULL to a
blank string.

Co-authored-by: JonMR <jon@radonfamily.com>