Fix default expression tests
There are some confusing edge cases in how ActiveRecord handles string
literals passed to the `change_column_default` method. We were currently
only catching one of the two edge cases and raising an error (when the
string literal contains an expression and surprisingly gets cast to
`nil` by ActiveRecord). That only currently happens in 5.2, but our test
was expecting it to happen in 5.0 and 5.1 also. So we need to limit the
test to only run when the edge case is present.

There's another edge case: in <= 5.1, a string literal containing an
expression is unexpectedly quoted by ActiveRecord resulting in the
expression being executed only once (at the time the DDL is executed)
rather than setting the expression itself as the default value to be
evaluated at runtime for every row insertion. We wrote a test to have
our migration method determine when you're encountering this edge case,
but it turns out that determining the edge case is in play is difficult
(without resorting to, for example, relying on the PgQuery gem, which we
don't believe is an acceptable runtime dependency for this gem, since it
would require people to have the Postgres dev packages installed.) So we
kept the test for now (hoping that in the future  we can determine a
reasonable way to catch this edge case) but pended it to prevent build
failures.

Co-authored-by: James Coleman <jtc331@gmail.com>