Check for default expressions unexpectedly evaluated at DDL time
Previously we skipped implementing this half of our default value
ambiguity checking because it appeared that the only real way to verify
the result wasn't surprising would be to parse the string value. But
after some more thought it turns out we can handle (virtually?) all
cases with a few heuristics. Effectively for non-string-ish types we can
disallow any string-typed values and require specifically typed objects.

In theory there's a possiblity that some custom object could be
coerced-Ruby side into a SQL string that does something weird here, but
I think that's an odd enough case that we can ignore it.