Report long transactions without  tables locked
Long-running transactions can block some kinds of DDL (at least from
completing) even if they haven't yet locked any tables. For example,
creating an index concurrently has to wait for all in-flight
transactions to complete before its second phase can run.

Along the way I also modified the reporting to show when a transaction
is idle or now and clarify if the query is currently running or is the
most-recently-run query for an idle transaction.