Fix: allow all string expression for enum-typed column defaults
After incorporating our recent version bump 1.2.0 into an app I realized
that enums are another special case to handle for our heuristics around
whether a default value will be parsed as an expression or as a
constant. They're virtually always represented as strings in SQL, but we
don't want to have to treat them as a special non-string type in Ruby,
so we need to allow string values by default.