Ensure unsafe methods are defined before SafeStatements are loaded
Prior to this change UnsafeStatements was being prepended after
SafeStatements. SafeStatements overrides several of the unsafe_* methods
which were then being blown away when UnsafeStatements was prepended.

In particular this meant that the SafeStatements implementation of
unsafe_add_column was not being executed, removing any guarantees about
the table being safely locked.