Raise helpful error for broken ActiveRecord index options handling
The combination of specifying `:opclass` in the `options` argument along
with a multi-column (or non-column expression) string passed as the
`column_names` argument results in ActiveRecord dropping the supplied
`:opclass` on the ground because `#index_column_names` returns the string
without processing and `#add_index_options` then calls
`#quoted_columns_for_index` with that string which then wraps that
string in an array _without processing_ (and adding per-column options).
(All of these calls are in `ActiveRecord::ConnectionAdapters::SchemaStatements`.)

Obviously that isn't great, so check for this situation and error so the
user doesn't get an unexpected result.