Make error classes inherit from StandardError not Exception
The error classes are named with the suffix `Error` anyway, so it's
confusing that they inherit from Ruby's `Exception` class. That's not a
nice surprise since `Exception` is conventionally special in the Ruby
world.

That being said we don't expect anyone to be rescuing these errors, but
it seems unnecessary to prevent it this way.