Add support for nulls_not_distinct option to index creation methods (#120)
This allows users to create indexes where NULL values are treated as equal
to each other rather than always being distinct, which can be useful for
certain uniqueness constraints.

Most of the `*_add_index` methods support this by default because we
already accept generic options, but in
`safe_add_concurrent_partitioned_index`  we need to support it specially
since we define kwargs instead of taking on `options` hash argument.

In all of the methods ActiveRecord defaults to simply dropping the
option on the floor when it is not supported by the database, which we
consider to be unsafe behavior, so we update all of the `safe_*` methods
to raise helpful errors when the required invariants don't hold.