Add idempotence and existing constraint re-use to safe_make_column_not_nullable
If an existing constraint already exists we should re-use that. If it's
not yet validated, we still need to validate it -- that gives use
idempotency, and we no longer need to raise an error if the previous run
failed. If the existing constraint is already validated, there's no need
to validate it, and we can simply proceed.

While it's true that _if you already know you have a validated
constraint_ we idiomatically prefer using
`safe_make_column_not_nullable_from_check_constraint` in cases where
your database state might be unknown (e.g., a previous failure), it's
still helpful to be able to use `safe_make_column_not_nullable`.

Fixes #130
Fixes #131